chore: mock 데이터 삭제 #87
