+++ frontend/src/components/Chart/BarCard/BarChart.tsx
@@ -1,9 +1,8 @@
 import * as S from './BarCard.style';
 import { useRef, useEffect, useState, useMemo } from 'react';
 import { Bar } from 'react-chartjs-2';
-import { monthFormatChanger, dayFormatChanger } from '../../../utils/dataFormatter';
-import { BarMonthItem, BarDayItem } from '../../../mocks/BarData';
-// import { useFilterStore } from '@/stores/filterStore';
+import { monthFormatChanger, dayFormatChanger } from '@/utils/dataFormatter';
+import { BarMonthItem, BarDayItem } from '@/types/chartType';
 import {
   Chart as ChartJS,
   ChartData,
