--- frontend/src/mocks/BarData.ts
@@ -1,359 +1,0 @@
-export interface BarMonthItem {
-  month: number;
-  fireCount: number;
-  assaultCount: number;
-  crowdCongestionCount: number;
-  weaponCount: number;
-  swoonCount: number;
-}
-
-export const BarMonthData: BarMonthItem[] = [
-  {
-    month: 1,
-    fireCount: 0,
-    assaultCount: 0,
-    crowdCongestionCount: 2,
-    weaponCount: 0,
-    swoonCount: 0,
-  },
-  {
-    month: 2,
-    fireCount: 5,
-    assaultCount: 4,
-    crowdCongestionCount: 1,
-    weaponCount: 0,
-    swoonCount: 2,
-  },
-  {
-    month: 3,
-    fireCount: 7,
-    assaultCount: 3,
-    crowdCongestionCount: 2,
-    weaponCount: 0,
-    swoonCount: 0,
-  },
-  {
-    month: 4,
-    fireCount: 3,
-    assaultCount: 7,
-    crowdCongestionCount: 3,
-    weaponCount: 2,
-    swoonCount: 1,
-  },
-  {
-    month: 5,
-    fireCount: 7,
-    assaultCount: 5,
-    crowdCongestionCount: 4,
-    weaponCount: 9,
-    swoonCount: 0,
-  },
-  {
-    month: 6,
-    fireCount: 5,
-    assaultCount: 2,
-    crowdCongestionCount: 2,
-    weaponCount: 3,
-    swoonCount: 1,
-  },
-  {
-    month: 7,
-    fireCount: 9,
-    assaultCount: 16,
-    crowdCongestionCount: 5,
-    weaponCount: 11,
-    swoonCount: 3,
-  },
-  {
-    month: 8,
-    fireCount:3,
-    assaultCount: 3,
-    crowdCongestionCount: 2,
-    weaponCount: 8,
-    swoonCount: 3,
-  },
-  {
-    month: 9,
-    fireCount: 11,
-    assaultCount: 9,
-    crowdCongestionCount: 12,
-    weaponCount: 5,
-    swoonCount: 2,
-  },
-  {
-    month: 10,
-    fireCount: 9,
-    assaultCount: 14,
-    crowdCongestionCount: 8,
-    weaponCount: 11,
-    swoonCount: 1,
-  },
-  {
-    month: 11,
-    fireCount: 8,
-    assaultCount: 2,
-    crowdCongestionCount: 1,
-    weaponCount: 13,
-    swoonCount: 9,
-  },
-  {
-    month: 12,
-    fireCount: 2,
-    assaultCount: 5,
-    crowdCongestionCount: 0,
-    weaponCount: 1,
-    swoonCount: 1,
-  },
-];
-
-
-export interface BarDayItem {
-  day: number;
-  fireCount: number;
-  assaultCount: number;
-  crowdCongestionCount: number;
-  weaponCount: number;
-  swoonCount: number;
-}
-export const BarDayData: BarDayItem[] = [
-  {
-    day: 1,
-    fireCount: 4,
-    assaultCount: 12,
-    crowdCongestionCount: 2,
-    weaponCount: 18,
-    swoonCount: 5,
-  },
-  {
-    day: 2,
-    fireCount: 5,
-    assaultCount: 13,
-    crowdCongestionCount: 2,
-    weaponCount: 19,
-    swoonCount: 13,
-  },
-  {
-    day: 3,
-    fireCount: 13,
-    assaultCount: 16,
-    crowdCongestionCount: 10,
-    weaponCount: 3,
-    swoonCount: 0,
-  },
-  {
-    day: 4,
-    fireCount: 15,
-    assaultCount: 1,
-    crowdCongestionCount: 2,
-    weaponCount: 7,
-    swoonCount: 5,
-  },
-  {
-    day: 5,
-    fireCount: 12,
-    assaultCount: 8,
-    crowdCongestionCount: 6,
-    weaponCount: 13,
-    swoonCount: 2,
-  },
-  {
-    day: 6,
-    fireCount: 2,
-    assaultCount: 19,
-    crowdCongestionCount: 7,
-    weaponCount: 6,
-    swoonCount: 15,
-  },
-  {
-    day: 7,
-    fireCount: 3,
-    assaultCount: 19,
-    crowdCongestionCount: 4,
-    weaponCount: 4,
-    swoonCount: 6,
-  },
-  {
-    day: 8,
-    fireCount: 20,
-    assaultCount: 4,
-    crowdCongestionCount: 12,
-    weaponCount: 10,
-    swoonCount: 20,
-  },
-  {
-    day: 9,
-    fireCount: 20,
-    assaultCount: 8,
-    crowdCongestionCount: 2,
-    weaponCount: 16,
-    swoonCount: 13,
-  },
-  {
-    day: 10,
-    fireCount: 18,
-    assaultCount: 0,
-    crowdCongestionCount: 10,
-    weaponCount: 19,
-    swoonCount: 19,
-  },
-  {
-    day: 11,
-    fireCount: 13,
-    assaultCount: 6,
-    crowdCongestionCount: 2,
-    weaponCount: 3,
-    swoonCount: 4,
-  },
-  {
-    day: 12,
-    fireCount: 4,
-    assaultCount: 2,
-    crowdCongestionCount: 13,
-    weaponCount: 6,
-    swoonCount: 12,
-  },
-  {
-    day: 13,
-    fireCount: 4,
-    assaultCount: 17,
-    crowdCongestionCount: 1,
-    weaponCount: 16,
-    swoonCount: 1,
-  },
-  {
-    day: 14,
-    fireCount: 7,
-    assaultCount: 2,
-    crowdCongestionCount: 17,
-    weaponCount: 19,
-    swoonCount: 16,
-  },
-  {
-    day: 15,
-    fireCount: 11,
-    assaultCount: 17,
-    crowdCongestionCount: 15,
-    weaponCount: 14,
-    swoonCount: 18,
-  },
-  {
-    day: 16,
-    fireCount: 8,
-    assaultCount: 19,
-    crowdCongestionCount: 16,
-    weaponCount: 17,
-    swoonCount: 19,
-  },
-  {
-    day: 17,
-    fireCount: 17,
-    assaultCount: 12,
-    crowdCongestionCount: 0,
-    weaponCount: 8,
-    swoonCount: 13,
-  },
-  {
-    day: 18,
-    fireCount: 2,
-    assaultCount: 17,
-    crowdCongestionCount: 6,
-    weaponCount: 4,
-    swoonCount: 17,
-  },
-  {
-    day: 19,
-    fireCount: 10,
-    assaultCount: 13,
-    crowdCongestionCount: 4,
-    weaponCount: 10,
-    swoonCount: 14,
-  },
-  {
-    day: 20,
-    fireCount: 18,
-    assaultCount: 2,
-    crowdCongestionCount: 1,
-    weaponCount: 11,
-    swoonCount: 7,
-  },
-  {
-    day: 21,
-    fireCount: 19,
-    assaultCount: 4,
-    crowdCongestionCount: 2,
-    weaponCount: 4,
-    swoonCount: 10,
-  },
-  {
-    day: 22,
-    fireCount: 10,
-    assaultCount: 18,
-    crowdCongestionCount: 4,
-    weaponCount: 16,
-    swoonCount: 18,
-  },
-  {
-    day: 23,
-    fireCount: 3,
-    assaultCount: 18,
-    crowdCongestionCount: 8,
-    weaponCount: 14,
-    swoonCount: 20,
-  },
-  {
-    day: 24,
-    fireCount: 3,
-    assaultCount: 11,
-    crowdCongestionCount: 6,
-    weaponCount: 6,
-    swoonCount: 3,
-  },
-  {
-    day: 25,
-    fireCount: 9,
-    assaultCount: 1,
-    crowdCongestionCount: 0,
-    weaponCount: 17,
-    swoonCount: 1,
-  },
-  {
-    day: 26,
-    fireCount: 11,
-    assaultCount: 19,
-    crowdCongestionCount: 18,
-    weaponCount: 6,
-    swoonCount: 13,
-  },
-  {
-    day: 27,
-    fireCount: 7,
-    assaultCount: 11,
-    crowdCongestionCount: 17,
-    weaponCount: 2,
-    swoonCount: 8,
-  },
-  {
-    day: 28,
-    fireCount: 1,
-    assaultCount: 15,
-    crowdCongestionCount: 17,
-    weaponCount: 14,
-    swoonCount: 16,
-  },
-  {
-    day: 29,
-    fireCount: 11,
-    assaultCount: 4,
-    crowdCongestionCount: 17,
-    weaponCount: 18,
-    swoonCount: 6,
-  },
-  {
-    day: 30,
-    fireCount: 15,
-    assaultCount: 6,
-    crowdCongestionCount: 12,
-    weaponCount: 20,
-    swoonCount: 6,
-  },
-];
