--- frontend/src/mocks/OverviewMock.ts
@@ -1,6 +1,0 @@
-import { overviewType } from '@/types/chartType';
-export const OVERVIEWMOCK: overviewType = {
-  recentCase: 15,
-  todayCase: 29,
-  mostCase: '화재',
-};
